+++ CHANGELOG.md
@@ -113,4 +113,5 @@ keep running if the sender is a player.
 * **Added** `ArgumentHandleResult#hasReason`
 * **Added** `DoubleArgument`
 * **Added** `DoubleArgumentNode`
-* **Added** `EnumArgument`: List arguments but for enums.
+* **Added** `EnumArgument`: List argument but for enums.
+* **Added** `EnumArgumentNode`: List argument but for enums.